120 change ux to be only use with and pickdrop

+++ ee_frontend/NDK-version/app/src/main/cpp/native-lib.c
@@ -95,7 +95,7 @@ int soundToPlay = -1;
 
 void stopSounds(void) {}
 
-void playSound(const int action) {
+void playSound(const uint8_t action) {
     soundToPlay = action;
 }
